services/user-service/requirements.txt
Django==5.0.1
djangorestframework-simplejwt==5.3.1
django-ninja==1.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
pandas==2.1.4
openpyxl==3.1.2
redis==5.0.1
requests==2.31.0
urllib3==2.1.0


